test(audio layer): add more tests
